PiratePanel/requirements.txt
customtkinter
Pillow
qrcode
beautifulsoup4
python-dotenv
requests